fix: Improve loading condition logic
In https://github.com/Automattic/BuddyPress-VIP-Go/pull/18, the undefined DOING_AJAX check was added, since none of the other negativised conditions are truthy in an Ajax request, so the filters weren't loading, and the BuddyBoss delete default cover image feature wasn't working.

However, this meant that as per https://github.com/Automattic/BuddyPress-VIP-Go/issues/22, the logic was running incorrectly - if the constant isn't defined (as per regular non-Ajax requests), then the function was hitting the `return`.

By refactoring this to remove the double-negatives, we can simplify the logic - if the Ajax constant is true, or the conditions for us working with the VIP Filesystem are true, then apply the filters.